--- drawing.py
@@ -1,8 +1,0 @@
-from settings import *
-
-import pygame
-import math
-
-
-class Drawing:
-    pass
