+++ main.py
@@ -1,8 +1,10 @@
 import sys
 
-from map import *
+from settings import *
+import pygame
 from mob.player import player
 from mob import mob
+from turf import turf
 
 
 class Game:
@@ -15,8 +17,9 @@ class Game:
 
     def new_game(self):
         self.mobhandler = mob.MobHandler(self)
-        self.mobhandler.add_mob(player.Player(self, (HALF_WIDTH, HALF_HEIGHT), icon="icon/mob/mob.png", icon_state="down"))
-        self.map = Map(self)
+        self.mobhandler.add_mob(player.Player(self, (HALF_WIDTH, HALF_HEIGHT),
+                                              icon="icon/mob/mob.png", icon_state="down"))
+        self.turfhandler = turf.TurfHandler(self)
 
     def update(self):
         pygame.display.flip()
@@ -25,8 +28,8 @@ class Game:
         self.mobhandler.process()
 
     def draw(self):
-        self.screen.fill('black')
-        self.map.draw()
+        self.screen.fill(BLACK)
+        self.turfhandler.draw()
         self.mobhandler.draw()
 
     def check_events(self):
