--- map.py
@@ -1,41 +1,0 @@
-from settings import *
-
-import pygame
-
-_ = False
-
-mini_map = [
-    [1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
-    [1, _, _, _, _, _, _, _, _, _, _, _, _, _, _, 1],
-    [1, _, _, 1, 1, 1, 1, _, _, _, 1, 1, 1, _, _, 1],
-    [1, _, _, _, _, _, 1, _, _, _, _, _, 1, _, _, 1],
-    [1, _, _, _, _, _, 1, _, _, _, _, _, 1, _, _, 1],
-    [1, _, _, 1, 1, 1, 1, _, _, _, _, _, _, _, _, 1],
-    [1, _, _, _, _, _, _, _, _, _, _, _, _, _, _, 1],
-    [1, _, _, _, 1, _, _, _, 1, _, _, _, _, _, _, 1],
-    [1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
-]
-
-
-class Map:
-    def __init__(self, game):
-        self.game = game
-        self.mini_map = mini_map
-        self.world_map = {}
-        self.get_map()
-
-    def get_map(self):
-        for j, row in enumerate(self.mini_map):
-            for i, value in enumerate(row):
-                if value:
-                    self.world_map[(i, j)] = value
-
-    def draw(self):
-        #[pygame.draw.rect(self.game.screen, DARKGRAY, (pos[0] * TILE, pos[1] * TILE, TILE, TILE), 2)
-        #  for pos in self.world_map]
-
-        [pygame.draw.rect(self.game.screen, DARKGRAY,
-                          (HALF_WIDTH - (self.game.mobhandler.get_player.x - pos[0] * TILE),
-                           HALF_HEIGHT - (self.game.mobhandler.get_player.y - pos[1] * TILE),
-                           TILE, TILE), 2)
-         for pos in self.world_map]
