+++ mob/mob.py
@@ -1,13 +1,13 @@
 from settings import *
 
-from main import Game
+import main
 import pygame
 import math
 import mob
 
 
 class Mob:
-    def __init__(self, game: Game, pos: tuple, icon: str = "", icon_state: str = "", health: int = 100, angle: int = 0):
+    def __init__(self, game, pos: tuple[int, int], icon: str = "", icon_state: str = "", health: int = 100, angle: int = 0):
         self.x, self.y = pos
         if icon:
             self.icon = pygame.image.load(icon).convert_alpha()
@@ -25,18 +25,18 @@ class Mob:
     def pos(self):
         return (self.x, self.y)
 
-    def draw(self, player_pos):
+    def draw(self, player_pos: tuple[int, int]):
         spr = self.get_sprite(0, 0)
         self.game.screen.blit(spr, spr.get_rect())
 
-    def get_sprite(self, x, y):
+    def get_sprite(self, x: int, y: int):
         sprite = pygame.Surface((TILE, TILE), pygame.SRCALPHA)
         sprite.blit(self.icon, (0, 0), (x * TILE, y * TILE, TILE, TILE))
         return sprite
 
 
 class MobHandler:
-    def __init__(self, game: Game):
+    def __init__(self, game):
         self.game = game
         self.mobs = list()
